+++ src/include/rsm-interface.h
@@ -18,6 +18,7 @@
 #ifndef RSM_INTERFACE_H
 #define RSM_INTERFACE_H
 #include <unistd.h>
+#include <stdint.h>
 
 typedef enum client_msg_code_t{
     C_SEND_WR = 0,
@@ -39,7 +40,7 @@ typedef struct client_msg_t{
 
 struct proxy_node_t;
 
-struct proxy_node_t* proxy_init(int node_id,const char* start_mode,const char* config_path,
+struct proxy_node_t* proxy_init(uint16_t node_id,const char* start_mode,const char* config_path,
         const char* log_path,int fake_mode);
 void proxy_run(struct proxy_node_t* proxy);
 void proxy_exit(struct proxy_node_t* proxy);
